backend/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-pptx==0.6.21
pandas==2.0.3
openpyxl==3.1.2
fpdf2==2.7.7
requests==2.31
passlib==1.7.4
jose==1.0.0
fastapi==0.115.12
stripe==12.1.0
jwt==1.3.1
python-dateutil==2.8.2
pytz==2023.3